V1.8 prerelease (#463)
* python3 port

* Delete .DS_Store

* a few updates

* handle signalP 5.0 parsing -- not tested yet

* more updates to support signalp 5.0 as well as input results at command line

* bump version and fix typo

* string format error in signalp

* ensure no empty sequences converting from bam to fasta

* more py2/3 fixes -- remove pybam in favor of samtools parsing

* DBxref to Dbxref in GFF3

* push some @photocyte PR changes to py3 dev branch, untested

* small doc fix in argparse usage

* try to fix unicode error in setupDB

* fix download in test

* update download print stmnt

* fix typo in predict

* fix genemark path call

* fix typo in exonerate_pident

* try to fix typos again

* fix p2g cmd

* fix genemark logic if not in path

* more messing with genemark stupid

* fix iprscan download for docker

* remove sam2bam.sh from library map

* fix typo in bam2gff3

* add busco_seed_species local directory check before running busco

* fix another type in bam parser

* remove .lib from library dummy

* add back augustus generic to local folder

* remove sam2bam.sh calls with subprocess.Popen

* add lowercase to snap; start to move downloads to json

* load download links from file

* add status msg to download json file

* fix a few typos in the bam alignment

* fix minimap2_cmd typo

* global FNULL

* fix indent error

* troubleshooting compare

* troubleshooting compare

* troubleshooting compare again

* troubleshooting compare yet again

* troubleshooting compare yet yet again

* troubleshooting compare yet yet again; try loc

* pandas bug fix, roll back debug print stmnts

* code linting

* update menu with p2g options

* typo - identidy to identity

* support explicit number of threads for tRNAscan-SE runs

* support explicit number of threads for tRNAscan-SE runs #411

* force to string for cpu arg

* fix trnascan cpus

* update dbCAN to v8

* filter cazy subdomains out if exist

* update EVM script to try to partition in between putative gene loci

* EVM update; PASA updated options; predict simplified busco

* fix typos in setup; add interpro fix

* incorporate py2 backwards compatibility; fix to interpo mapping terms

* updates to resources and typo fix

* update pip install instructions, no longer py2 only

* update setup menu with --local option

* fix --rename option in annotate

* add io.open to setup for py2 compatibility

* updateTBL for error catching

* updateTBL for error catching

* updateTBL for error catching 2

* add ncRNA parsing in tbl2dict

* test fix for antiSMASH v5 parsing

* support for ncRNA in gff output

* dynamic output of annotation table for custom headers

* unify gb and gff3 dictionary

* fix resources COGS bug

* dont let smcogs be their own column

* write synonyms/alias to tbl file

* preference to custom annotations over automated

* make sure synonyms are not name and do not repeat

* make sure synonyms func update

* make sure synonyms func update not working, print to debug

* clean func diff logic

* clean func diff logic

* clean func working, remove debug

* fix embarassing typo for cazymes

* troubleshoot cazy subdomain reduction

* fix cazy parsing

* exit on parse error for custom

* exit on parse error for custom

* fix tbl parser for rRNA features

* fix minimap2 bam2hints function, hopefully fix #444

* add alias= parsing for gff2dict

* just some formatting changes to code

* make sure synonyms are not duplcated in gff3 parsing

* do not enumerate gene names if from alt transcripts

* do not enumerate gene names if from alt transcripts

* add debug print stmnt to naming method

* hopfully fixed, forgot a len()

* annotation table error to terminal

* annotation table error to terminal

* annotation table fix for note field

* fix gff3 parsing skip all features it cant parse

* updates to support strange genbank GFF3 files

* fix for gbk parsing to capture misc_RNA as ncRNA

* fix typo on pasa DB name

* fix pasa function in update for string

* same str change in train

* try to fix gene number count if atypical locus_tag, #453

* minor update to latest fix

* minor update to latest fix

* update logging for GO enrichment in compare

* better error handling for subprocess calls

* fix py3 incompatibility in annotationtable

* if --organism=other then make codingquarry default off unless valid --weights passed

* specify the pasa conf file in env variable

* replace . from names as well in dbname

* require distro for linux_distribution info for python 3.8 and beyond

* remove merge comments for commandline options in update

+++ funannotate/test.py
@@ -4,11 +4,15 @@
 import sys
 import os
 import subprocess
-import urllib2
+try:
+    from urllib.request import urlopen
+except ImportError:
+    from urllib2 import urlopen
 import socket
 import argparse
 import shutil
 import errno
+
 
 def checkFile(input):
     def _getSize(filename):
@@ -28,7 +32,7 @@ def checkFile(input):
 
 def countfasta(input):
     count = 0
-    with open(input, 'rU') as f:
+    with open(input, 'r') as f:
         for line in f:
             if line.startswith(">"):
                 count += 1
@@ -37,7 +41,7 @@ def countfasta(input):
 
 def countGFFgenes(input):
     count = 0
-    with open(input, 'rU') as f:
+    with open(input, 'r') as f:
         for line in f:
             if "\tgene\t" in line:
                 count += 1
@@ -45,7 +49,7 @@ def countGFFgenes(input):
 
 
 def runCMD(cmd, dir):
-    print('CMD: {:}'.format(' '.join(cmd)))
+    print(('CMD: {:}'.format(' '.join(cmd))))
     print("#########################################################")
     subprocess.call(cmd, cwd=dir)
 
@@ -53,11 +57,14 @@ def runCMD(cmd, dir):
 def download(url, name):
     file_name = name
     try:
-        u = urllib2.urlopen(url)
+        u = urlopen(url)
         f = open(file_name, 'wb')
         meta = u.info()
-        file_size = int(meta.getheaders("Content-Length")[0])
-        print("Downloading: {0} Bytes: {1}".format(url, file_size))
+        file_size = 0
+        for x in meta.items():
+            if x[0].lower() == 'content-length':
+                file_size = int(x[1])
+        print("Downloading: {} Bytes: {}".format(url, file_size))
         file_size_dl = 0
         block_sz = 8192
         while True:
@@ -67,7 +74,7 @@ def download(url, name):
             file_size_dl += len(buffer)
             f.write(buffer)
             p = float(file_size_dl) / file_size
-            status = r"{0}  [{1:.2%}]".format(file_size_dl, p)
+            status = r"{}  [{:.2%}]".format(file_size_dl, p)
             status = status + chr(8)*(len(status)+1)
             sys.stdout.write(status)
         sys.stdout.flush()
@@ -121,7 +128,8 @@ def runCleanTest(args):
     try:
         assert countfasta(os.path.join(tmpdir, 'test.exhaustive.fa')) == 3
         print('SUCCESS: `funannotate clean` test complete.')
-        shutil.rmtree(tmpdir)
+        if not args.debug:
+            shutil.rmtree(tmpdir)
     except AssertionError:
         print('ERROR: `funannotate clean` test failed.')
     print("#########################################################\n")
@@ -151,7 +159,8 @@ def runPredictTest(args):
         assert 1500 <= countGFFgenes(os.path.join(
             tmpdir, 'annotate', 'predict_results', 'Awesome_testicus.gff3')) <= 1800
         print('SUCCESS: `funannotate predict` test complete.')
-        shutil.rmtree(tmpdir)
+        if not args.debug:
+            shutil.rmtree(tmpdir)
     except AssertionError:
         print('ERROR: `funannotate predict` test failed - check logfiles')
     print("#########################################################\n")
@@ -191,10 +200,10 @@ def runBuscoTest(args):
             tmpdir, 'annotate', 'predict_results', 'Awesome_busco.gff3')) <= 1800
         print('SUCCESS: `funannotate predict` BUSCO-mediated training test complete.')
         print("#########################################################")
-        #print('Adding training parameters to database')
+        # print('Adding training parameters to database')
         # now lets try to add it to the database as new species
-        #cmd2 = ['funannotate', 'species', '-s', 'awesome_busco', '-p', 'annotate/predict_results/awesome_busco.parameters.json']
-        #runCMD(cmd2, tmpdir)
+        # cmd2 = ['funannotate', 'species', '-s', 'awesome_busco', '-p', 'annotate/predict_results/awesome_busco.parameters.json']
+        # runCMD(cmd2, tmpdir)
         # now lets try to run this again, using the parameters
         # print("#########################################################")
         print('Now running predict using all pre-trained ab-initio predictors')
@@ -209,12 +218,13 @@ def runBuscoTest(args):
                 tmpdir, 'annotate2', 'predict_results', 'Awesome_busco.gff3')) <= 1800
             print(
                 'SUCCESS: `funannotate predict` using existing parameters test complete.')
-            shutil.rmtree(tmpdir)
+            if not args.debug:
+                shutil.rmtree(tmpdir)
         except AssertionError:
             print(
                 'ERROR: `funannotate predict` using existing parameters test failed - check logfiles')
         # delete the training data from database
-        #shutil.rmtree(os.path.join(FUNDB, 'trained_species', 'awesome_busco'))
+        # shutil.rmtree(os.path.join(FUNDB, 'trained_species', 'awesome_busco'))
     except AssertionError:
         print('ERROR: `funannotate predict` BUSCO-mediated training test failed - check logfiles')
     print("#########################################################\n")
@@ -254,7 +264,8 @@ def runAnnotateTest(args):
         assert checkFile(os.path.join(tmpdir, 'annotate',
                                       'annotate_results', 'Genome_one.annotations.txt'))
         print('SUCCESS: `funannotate annotate` test complete.')
-        shutil.rmtree(tmpdir)
+        if not args.debug:
+            shutil.rmtree(tmpdir)
     except AssertionError:
         print('ERROR: `funannotate annotate` test failed - check logfiles')
     print("#########################################################\n")
@@ -287,7 +298,8 @@ def runCompareTest(args):
         assert checkFile(os.path.join(tmpdir, 'compare', 'phylogeny.html'))
         assert checkFile(os.path.join(tmpdir, 'compare.tar.gz'))
         print('SUCCESS: `funannotate compare` test complete.')
-        shutil.rmtree(tmpdir)
+        if not args.debug:
+            shutil.rmtree(tmpdir)
     except AssertionError:
         print('ERROR: `funannotate compare` test failed - check logfiles')
     print("#########################################################\n")
@@ -344,7 +356,8 @@ def runRNAseqTest(args):
             assert 1630 <= countGFFgenes(os.path.join(
                 tmpdir, 'rna-seq', 'update_results', 'Awesome_rna.gff3')) <= 1830
             print('SUCCESS: funannotate RNA-seq training/prediction test complete.')
-            shutil.rmtree(tmpdir)
+            if not args.debug:
+                shutil.rmtree(tmpdir)
         except AssertionError:
             print(
                 'ERROR: funannotate RNA-seq training/prediction test failed - check logfiles')
@@ -367,6 +380,8 @@ def main(args):
                         choices=['all', 'clean', 'mask', 'predict',
                                  'busco', 'annotate', 'rna-seq', 'compare'],
                         help='select which tests to run')
+    parser.add_argument('--debug', action='store_true',
+                        help='keep folder')
     parser.add_argument('--cpus', default=2, type=int,
                         help='Number of CPUs to use')
     args = parser.parse_args(args)
